requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
PyJWT==2.8.0
httpie==3.2.2
Flask-Mail==0.10.0
